+++ front/app/containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/messages.ts
@@ -78,22 +78,6 @@ export default defineMessages({
     id: 'app.containers.AdminPage.SettingsPage.imageOverlayOpacity',
     defaultMessage: 'Image overlay opacity',
   },
-  bgHeaderPreviewSelectLabel: {
-    id: 'app.containers.AdminPage.SettingsPage.bgHeaderPreviewSelectLabel',
-    defaultMessage: 'Show preview for',
-  },
-  desktop: {
-    id: 'app.containers.AdminPage.SettingsPage.desktop',
-    defaultMessage: 'Desktop',
-  },
-  tablet: {
-    id: 'app.containers.AdminPage.SettingsPage.tablet',
-    defaultMessage: 'Tablet',
-  },
-  phone: {
-    id: 'app.containers.AdminPage.SettingsPage.phone',
-    defaultMessage: 'Phone',
-  },
   noHeader: {
     id: 'app.containers.AdminPage.SettingsPage.noHeader',
     defaultMessage: 'Please upload a header image',